requirements.txt
# Core dependencies
geopy==2.4.1
pandas==2.2.0
requests==2.31.0
beautifulsoup4==4.12.3
folium>=0.19.0
pillow==10.2.0
ipywidgets==8.1.2
streamlit==1.31.1
transformers==4.37.2
torch==2.2.0
fpdf==1.7.2
jupyter==1.0.0
numpy==1.26.4
python-dotenv==1.0.1
streamlit-folium
plotly==5.18.0

# API and HTTP
urllib3==2.2.0
aiohttp==3.9.3
httpx==0.26.0

# Data validation and processing
pydantic==2.6.1
jsonschema==4.21.1
email-validator==2.1.0.post1

# Caching and performance
cachetools==5.3.2
redis==5.0.1
aioredis==2.0.1

# Testing and development
pytest==8.0.0
pytest-cov==4.1.0
pytest-asyncio==0.23.5
black==24.1.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0

# Logging and monitoring
structlog==24.1.0
sentry-sdk==1.39.1

# Security
cryptography==42.0.2
python-jose==3.3.0
passlib==1.7.4

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==2.0.0
 